requirements.txt
django==4.2.28
django-extensions==4.1
django-debug-toolbar==6.2.0
